src/main/java/site/zoraiarodrigues/spring/repository/RequestRepository.java
package site.zoraiarodrigues.spring.repository;

import java.util.List;

import org.springframework.data.jpa.repository.JpaRepository;
import org.springframework.data.jpa.repository.Query;
import org.springframework.stereotype.Repository;

import site.zoraiarodrigues.spring.domain.Request;
import site.zoraiarodrigues.spring.domain.enums.RequestState;

@Repository
public interface RequestRepository extends JpaRepository<Request, Long> {
	
	public List<Request> findAllByOwnerId(Long id);
	
	@Query("UPDATE request SET state = ?2 WHERE id = ?1")
	public Request updateRequest(Long id, RequestState state);
	
}
